Move tracing.LogWithLevel to logger.LogWithLevel
